requirements.txt
pytest
fastapi
email-validator
python-dotenv
requests
pydantic
fastapi-pagination
uvicorn
faker
